きちい

+++ Engine/Base/Vertex.cpp
@@ -6,13 +6,13 @@ bool VertexPNU::operator==(const VertexPNU& a) const
 	return pos == a.pos && normal == a.normal && uv == a.uv;
 }
 
-void VertexPNU::CalcNormalVec(VertexPNU list[], const unsigned int indexlist[], const unsigned int indexcount)
+void VertexPNU::CalcNormalVec(VertexPNU list[], const uint32_t indexlist[], const uint32_t indexcount)
 {
 	assert(indexcount % 3 == 0);
-	for (unsigned int i = 0; i < indexcount / 3; i++) {
-		unsigned int index0 = indexlist[i * 3 + 0];
-		unsigned int index1 = indexlist[i * 3 + 1];
-		unsigned int index2 = indexlist[i * 3 + 2];
+	for (uint32_t i = 0; i < indexcount / 3; i++) {
+		uint32_t index0 = indexlist[i * 3 + 0];
+		uint32_t index1 = indexlist[i * 3 + 1];
+		uint32_t index2 = indexlist[i * 3 + 2];
 
 		Vector3 p0 = list[index0].pos;
 		Vector3 p1 = list[index1].pos;
@@ -30,12 +30,12 @@ void VertexPNU::CalcNormalVec(VertexPNU list[], const unsigned int indexlist[], 
 	}
 }
 
-void VertexPNU::CalcNormalVec(std::vector<VertexPNU> list, std::vector<unsigned int> indexlist) {
+void VertexPNU::CalcNormalVec(std::vector<VertexPNU> list, std::vector<uint32_t> indexlist) {
 	assert(indexlist.size() % 3 == 0);
-	for (unsigned int i = 0; i < list.size() / 3; i++) {
-		unsigned int index0 = indexlist[i * 3 + 0];
-		unsigned int index1 = indexlist[i * 3 + 1];
-		unsigned int index2 = indexlist[i * 3 + 2];
+	for (size_t i = 0; i < list.size() / 3; i++) {
+		size_t index0 = indexlist[i * 3 + 0];
+		size_t index1 = indexlist[i * 3 + 1];
+		size_t index2 = indexlist[i * 3 + 2];
 
 		Vector3 p0 = list[index0].pos;
 		Vector3 p1 = list[index1].pos;
